Fixed the umulh and smulh instruction

+++ tests/lifter/smulh.rs
@@ -9,7 +9,7 @@ fn test_smulh_1() {
         check: // entry block
         nextln: v37 = i64.read_reg "x2"
         nextln: v38 = i64.read_reg "x3"
-        nextln: v39 = i128.imul v37, v38
+        nextln: v39 = i64.imul v37, v38
         nextln: v40 = i128.ashr v39, 0x40
         nextln: i64.write_reg v40, "x1"
     "#;
